+++ admin/e_bandobas/lib/app/modules/dutypoint/views/dutypoint_view.dart
@@ -1,4 +1,3 @@
-import 'package:e_bandobas/app/modules/dutypoint/views/pointview.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/src/widgets/framework.dart';
 import 'package:get/get.dart';
