+++ admin/e_bandobas/lib/app/modules/event/controllers/event_controller.dart
@@ -12,13 +12,14 @@ class EventController extends GetxController {
   void onInit() {
     super.onInit();
     loadEvents();
+    update();
   }
   
   void loadEvents() async {
     events.value = await EventApi.obtainEvents(API_Decision.Only_Failure);
     if (events.value != null && events.value!.isNotEmpty) {
-      selectedEventId.value = events.value!.elementAt(0).id!.toInt();
     }
+    events.refresh();
     update();
   }
   void saveEvent() {}
